refactoring almost was done

+++ lib/models/post_creator/block_photos.dart
@@ -6,12 +6,13 @@ import 'block_post.dart';
 class BlockPhotos extends BlockPost {
   List<String> paths;
   MethodViewPhoto methodView;
+  List<int> photoSizes;
   static const limitPhotos = 10;
 
   bool get isFull => paths.length >= limitPhotos;
 
-  BlockPhotos({List<String>? paths, this.methodView = MethodViewPhoto.tiles})
-      : paths = paths ?? [],
+  BlockPhotos({List<String>? paths, List<int>? photoSizes, this.methodView = MethodViewPhoto.tiles})
+      : paths = paths ?? [], photoSizes = photoSizes ?? [],
         super(type: BlockPostType.photos);
 
   @override
@@ -22,31 +23,48 @@ class BlockPhotos extends BlockPost {
     return false;
   }
 
-  bool addPath(String path) {
+  Future<bool> addPath(String path) async {
     if (paths.length >= limitPhotos) {
       return false;
     }
-    paths.add(path);
-    return true;
+
+    try {
+      final file = File(path);
+      if (await file.exists()) {
+        int size = await file.length();
+        paths.add(path);
+        photoSizes.add(size); // Добавляем размер в соответствующий индекс
+        return true;
+      }
+    } catch (e) {
+      print("Ошибка при получении размера файла: $e");
+    }
+    return false;
   }
 
   Future<void> deletePhoto(int index) async {
-    if (await File(paths[index]).exists()) await File(paths[index]).delete();
+    try {
+      if (await File(paths[index]).exists()) {
+        await File(paths[index]).delete();
+      }
+    } catch (e) {
+      print("Файл уже удален или недоступен");
+    }
+
     paths.removeAt(index);
+    photoSizes.removeAt(index); // Удаляем размер вместе с путем
   }
 
   Future<void> deleteAllPhotos() async {
     for (String path in paths) {
       try {
         final file = File(path);
-        if (await file.exists()) {
-          await file.delete();
-        }
+        if (await file.exists()) await file.delete();
       } catch (e) {
-        // Логируем ошибку, если файл занят другим процессом
         print("Ошибка при удалении файла $path: $e");
       }
     }
-
+    paths.clear();
+    photoSizes.clear();
   }
 }
